backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
kafka-python==2.0.2
aioredis==2.0.1
websockets==12.0
aiohttp==3.9.1
email-validator==2.1.0
